chore: move timestats to root internal package
