+++ runner/internal/metrics/aggregate.go
@@ -3,8 +3,8 @@ package metrics
 import (
 	"time"
 
+	"github.com/benchttp/engine/internal/timestats"
 	"github.com/benchttp/engine/runner/internal/recorder"
-	"github.com/benchttp/engine/runner/internal/timestats"
 )
 
 // Aggregate is an aggregate of metrics computed from
