+++ runner/internal/report/report_test.go
@@ -5,11 +5,11 @@ import (
 	"time"
 
 	"github.com/benchttp/engine/internal/cli/ansi"
+	"github.com/benchttp/engine/internal/timestats"
 	"github.com/benchttp/engine/runner/internal/config"
 	"github.com/benchttp/engine/runner/internal/metrics"
 	"github.com/benchttp/engine/runner/internal/report"
 	"github.com/benchttp/engine/runner/internal/tests"
-	"github.com/benchttp/engine/runner/internal/timestats"
 )
 
 func TestReport_String(t *testing.T) {
