+++ runner/internal/tests/tests_test.go
@@ -5,9 +5,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/benchttp/engine/internal/timestats"
 	"github.com/benchttp/engine/runner/internal/metrics"
 	"github.com/benchttp/engine/runner/internal/tests"
-	"github.com/benchttp/engine/runner/internal/timestats"
 )
 
 func TestRun(t *testing.T) {
